Push files

+++ Mat3am Elhabaib/Models/OrderItems.cs
@@ -10,8 +10,8 @@ namespace Mat3am_Elhabaib.Models
         public int Id { get; set; }
 
         public int? Quantity { get; set; }
-        public double? UnitPrice { get; set; }
-        public double? TotalPrice { get; set; }
+        public decimal? UnitPrice { get; set; }
+        public decimal? TotalPrice { get; set; }
 
         public int ItemId { get; set; }
         [ForeignKey(nameof(ItemId))]
